api/requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
chardet==3.0.4
Django==2.1.2
djangorestframework==3.9.0
djangorestframework-gis==0.13
gunicorn==19.9.0
google-auth==1.6.0
geopy==1.18.0
idna==2.7
kombu==4.2.1
psycopg2-binary==2.7.6.1
pytz==2018.7
redis==2.10.6
requests==2.20.0
urllib3==1.24
vine==1.1.4
GDAL==2.1.0
Pillow==5.4.1
